+++ src/app/layout.tsx
@@ -1,6 +1,16 @@
 import type { Metadata } from "next";
 import { backdrop } from "@/lib/class-utils";
-import PageLayout from "@/components/page-layout";
+
+import "@/app/globals.css";
+import { Inter } from "next/font/google";
+import { GoogleTagManager } from "@next/third-parties/google";
+import { PropsWithChildren } from "react";
+const inter = Inter({
+  subsets: ["latin"],
+  display: "swap",
+  variable: "--font-inter",
+  weight: "600",
+});
 
 export const metadata: Metadata = {
   title: "Matthew Simo",
@@ -14,8 +24,14 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <PageLayout>
-      <main className={backdrop}>{children}</main>
-    </PageLayout>
+    <html lang="en">
+      <head>
+        <link rel="me" href="https://mastodon.social/@matthewsimo" />
+      </head>
+      <GoogleTagManager gtmId="G-4SBJQ1ZHYZ" />
+      <body className={`${inter.variable} font-sans text-main bg-main`}>
+        {children}
+      </body>
+    </html>
   );
 }
