+++ src/app/page.tsx
@@ -1,25 +1,27 @@
-import { gridClass } from "@/lib/class-utils";
 import Link from "@/components/link";
+import { backdrop, gridClass } from "@/lib/class-utils";
 
 export default function Home() {
   return (
-    <div className={`${gridClass} leading-relaxed text-xl space-y-8 my-40`}>
-      <span className="text-3xl">Howdy</span>
-      <h1 className="text-5xl">I&apos;m Matthew&nbsp;Simo</h1>
+    <main className={backdrop}>
+      <div className={`${gridClass} leading-relaxed text-xl space-y-8 my-40`}>
+        <span className="text-3xl">Howdy</span>
+        <h1 className="text-5xl">I&apos;m Matthew&nbsp;Simo</h1>
 
-      <p>
-        I&apos;m a Software Engineer & UX Designer based in Texas and I like to
-        build things for humans. I work as a Senior Solutions Architect of
-        Front-End Engineering at{" "}
-        <Link href="https://www.rightpoint.com/">Rightpoint</Link> where I lead
-        teams that build software that aims to improves people&rsquo;s lives.
-        You can usually find me exploring the intersection of humans &amp;
-        computers.
-      </p>
-      <p>
-        <Link href="/about">Find out more</Link> or{" "}
-        <Link href="/posts">see some posts</Link>
-      </p>
-    </div>
+        <p>
+          I&apos;m a Software Engineer & UX Designer based in Texas and I like
+          to build things for humans. I work as a Senior Solutions Architect of
+          Front-End Engineering at{" "}
+          <Link href="https://www.rightpoint.com/">Rightpoint</Link> where I
+          lead teams that build software that aims to improves people&rsquo;s
+          lives. You can usually find me exploring the intersection of humans
+          &amp; computers.
+        </p>
+        <p>
+          <Link href="/about">Find out more</Link> or{" "}
+          <Link href="/posts">see some posts</Link>
+        </p>
+      </div>
+    </main>
   );
 }
